@@ -1,158 +1,150 @@
 {
   "name": "@lyrio/ui",
-  "version": "0.2.7",
+  "version": "0.3.4",
   "devDependencies": {
-    "@fontsource/dm-mono": "^4.0.0",
-    "@fontsource/fira-code": "^4.0.0",
-    "@fontsource/inconsolata": "^4.0.0",
-    "@fontsource/jetbrains-mono": "^4.0.0",
-    "@fontsource/noto-sans": "^4.0.0",
-    "@fontsource/noto-serif": "^4.0.0",
-    "@fontsource/nunito-sans": "^4.0.0",
-    "@fontsource/open-sans": "^4.0.0",
-    "@fontsource/pt-mono": "^4.0.0",
-    "@fontsource/pt-sans": "^4.0.0",
-    "@fontsource/roboto": "^4.0.0",
-    "@fontsource/roboto-mono": "^4.0.0",
-    "@fontsource/roboto-slab": "^4.0.0",
-    "@fontsource/source-code-pro": "^4.0.0",
-    "@fontsource/source-serif-pro": "^4.0.0",
-    "@fontsource/ubuntu-mono": "^4.0.0",
-    "@fontsource/zilla-slab": "^4.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/plugin-proposal-decorators": "^7.17.2",
+    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
+    "@esm-bundle/markdown-it": "^12.3.2",
+    "@fontsource/dm-mono": "^4.5.2",
+    "@fontsource/fira-code": "^4.5.5",
+    "@fontsource/inconsolata": "^4.5.2",
+    "@fontsource/jetbrains-mono": "^4.5.3",
+    "@fontsource/noto-sans": "^4.5.3",
+    "@fontsource/noto-serif": "^4.5.3",
+    "@fontsource/nunito-sans": "^4.5.2",
+    "@fontsource/open-sans": "^4.5.5",
+    "@fontsource/pt-mono": "^4.5.1",
+    "@fontsource/pt-sans": "^4.5.1",
+    "@fontsource/roboto": "^4.5.3",
+    "@fontsource/roboto-mono": "^4.5.3",
+    "@fontsource/roboto-slab": "^4.5.4",
+    "@fontsource/source-code-pro": "^4.5.4",
+    "@fontsource/source-serif-pro": "^4.5.2",
+    "@fontsource/ubuntu-mono": "^4.5.2",
+    "@fontsource/zilla-slab": "^4.5.1",
     "@hackmd/markdown-it-task-lists": "^2.1.1",
-    "@tanem/react-nprogress": "^3.0.53",
-    "@types/blueimp-md5": "^2.7.0",
-    "@types/chart.js": "^2.9.29",
-    "@types/css": "^0.0.31",
-    "@types/flat": "^5.0.1",
-    "@types/gravatar": "^1.8.0",
-    "@types/js-yaml": "^4.0.0",
-    "@types/lodash.clonedeep": "^4.5.6",
-    "@types/lodash.isequal": "^4.5.5",
-    "@types/lodash.merge": "^4.6.6",
-    "@types/markdown-it": "^12.0.1",
-    "@types/node": "^14.14.22",
-    "@types/object-path": "^0.11.0",
-    "@types/prismjs": "^1.16.2",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/socket.io-client": "^1.4.35",
-    "@types/streamsaver": "^2.0.0",
-    "@types/twemoji": "^12.1.1",
-    "@types/ua-parser-js": "^0.7.35",
-    "@types/uuid": "^8.3.0",
-    "@types/validator": "^13.1.3",
-    "@types/webpack-env": "^1.16.0",
-    "ansi_up": "^4.0.4",
+    "@tanem/react-nprogress": "^4.0.7",
+    "@types/blueimp-md5": "^2.18.0",
+    "@types/chart.js": "^2.9.35",
+    "@types/css": "^0.0.33",
+    "@types/flat": "^5.0.2",
+    "@types/gravatar": "^1.8.3",
+    "@types/html-minifier-terser": "^6.1.0",
+    "@types/js-yaml": "^4.0.5",
+    "@types/lodash": "^4.14.178",
+    "@types/markdown-it": "^12.2.3",
+    "@types/node": "16",
+    "@types/object-path": "^0.11.1",
+    "@types/prismjs": "^1.26.0",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/serialize-javascript": "^5.0.2",
+    "@types/streamsaver": "^2.0.1",
+    "@types/svgo": "^2.6.2",
+    "@types/twemoji": "^12.1.2",
+    "@types/ua-parser-js": "^0.7.36",
+    "@types/uuid": "^8.3.4",
+    "@types/validator": "^13.7.1",
+    "@vitejs/plugin-legacy": "^1.7.1",
+    "@vitejs/plugin-react": "^1.2.0",
+    "ansi_up": "^5.1.0",
     "autotrack": "^2.4.1",
-    "axios": "^0.21.1",
+    "axios": "^0.26.0",
     "babel-plugin-import": "^1.13.3",
-    "babel-plugin-prismjs": "^2.0.1",
-    "blueimp-md5": "^2.18.0",
-    "chart.js": "^2.9.4",
-    "class-validator": "^0.13.1",
+    "babel-plugin-prismjs": "^2.1.0",
+    "blueimp-md5": "^2.19.0",
+    "browserslist-generator": "^1.0.65",
+    "chart.js": "^3.7.1",
+    "class-validator": "^0.13.2",
     "css": "^3.0.0",
     "css-element-queries": "^1.2.3",
-    "css-url-relative-plugin": "^1.0.1",
+    "css-url-relative-plugin": "^1.1.0",
     "customize-cra": "^1.0.0",
     "cz-conventional-changelog": "3.3.0",
-    "dayjs": "^1.10.3",
+    "dayjs": "^1.10.7",
     "dtsgenerator": "2.7.0",
-    "extract-loader": "^5.1.0",
-    "file-loader": "^6.2.0",
     "flat": "^5.0.2",
-    "fomantic-ui-css": "^2.8.7",
-    "fs-extra": "^9.1.0",
+    "fomantic-ui-css": "^2.8.8",
+    "fs-extra": "^10.0.0",
     "git-repo-info": "^2.1.1",
-    "gravatar": "^1.8.1",
+    "gravatar": "^1.8.2",
     "hack-font": "source-foundry/Hack#1eddf94",
+    "husky": "^7.0.4",
     "immutability-helper": "^3.1.1",
-    "js-yaml": "^4.0.0",
+    "js-yaml": "^4.1.0",
     "jsondiffpatch": "^0.4.1",
-    "less": "^4.1.0",
-    "less-loader": "^7.3.0",
-    "lodash.clonedeep": "^4.5.0",
-    "lodash.isequal": "^4.5.0",
-    "lodash.merge": "^4.6.2",
-    "markdown-it": "^12.0.4",
+    "less": "^4.1.2",
+    "lodash-es": "^4.17.21",
     "markdown-it-math-loose": "^4.1.1",
     "markdown-it-mentions": "^1.0.0",
     "markdown-it-merge-cells": "^1.0.1",
-    "mathjax-full": "^3.1.2",
-    "mobx": "^4.15.4",
+    "mathjax-full": "^3.2.0",
+    "mobx": "^6.3.13",
     "mobx-persist": "^0.4.1",
-    "mobx-react-lite": "^2.2.2",
-    "mobx-utils": "^5.6.2",
+    "mobx-react-lite": "^3.2.3",
+    "mobx-utils": "^6.0.4",
     "monaco-ace-tokenizer": "^0.2.1",
-    "monaco-editor": "0.21.2",
-    "monaco-editor-webpack-plugin": "2.0.0",
-    "monaco-themes": "^0.3.3",
+    "monaco-editor": "0.32.1",
+    "monaco-themes": "^0.4.0",
     "navi": "^0.14.0",
-    "noty": "^3.2.0-beta",
-    "object-path": "^0.11.5",
-    "p-all": "^3.0.0",
-    "patch-package": "^6.2.2",
+    "node-html-parser": "^5.2.0",
+    "noty": "^3.2.0-beta-deprecated",
+    "object-path": "^0.11.8",
+    "p-all": "^4.0.0",
+    "patch-package": "^6.4.7",
     "postinstall-postinstall": "^2.1.0",
-    "prettier": "^2.2.1",
-    "prismjs": "^1.23.0",
-    "react": "^17.0.0",
-    "react-app-rewired": "^2.1.8",
-    "react-chartjs-2": "2.11.1",
-    "react-countdown": "^2.3.1",
-    "react-dom": "^17.0.0",
-    "react-google-recaptcha-v3": "^1.7.1",
-    "react-helmet-async": "^1.0.7",
-    "react-intl": "^5.10.16",
-    "react-monaco-editor": "0.40.0",
+    "prettier": "^2.5.1",
+    "prismjs": "^1.26.0",
+    "react": "^17.0.2",
+    "react-chartjs-2": "4.0.1",
+    "react-countdown": "^2.3.2",
+    "react-dom": "^17.0.2",
+    "react-google-recaptcha-v3": "^1.9.7",
+    "react-helmet-async": "^1.2.2",
+    "react-intl": "^5.24.6",
+    "react-monaco-editor": "0.47.0",
     "react-navi": "^0.14.4",
-    "react-scripts": "4.0.1",
-    "react-textarea-autosize": "^8.3.0",
-    "semantic-ui-react": "^2.0.3",
+    "react-textarea-autosize": "^8.3.3",
+    "rollup-plugin-node-polyfills": "^0.2.1",
+    "semantic-ui-react": "^2.1.1",
+    "serialize-javascript": "^6.0.0",
     "sf-mono-webfont": "^1.0.0",
-    "socket.io-client": "^2.3.0",
-    "socket.io-msgpack-parser": "^2.2.1",
-    "streamsaver": "^2.0.5",
-    "string-replace-to-array": "^1.0.3",
-    "svgo": "^1.3.2",
-    "svgo-loader": "^2.2.1",
+    "socket.io-client": "^4.4.1",
+    "socket.io-msgpack-parser": "^3.0.1",
+    "streamsaver": "^2.0.6",
+    "string-replace-to-array": "^2.1.0",
+    "svgo": "^2.8.0",
     "timeago.js": "^4.0.2",
-    "twemoji": "^13.0.1",
+    "twemoji": "^13.1.0",
     "typeface-lato": "^1.1.13",
     "typeface-saira": "^1.1.13",
-    "typescript": "~4.1.3",
-    "ua-parser-js": "^0.7.23",
-    "use-debounce": "^5.2.0",
+    "typescript": "~4.5.5",
+    "ua-parser-js": "^1.0.2",
+    "use-debounce": "^7.0.1",
     "uuid": "^8.3.2",
-    "val-loader": "^2.1.2",
-    "validator": "^13.5.2",
+    "validator": "^13.7.0",
+    "vite": "^2.8.1",
+    "vite-plugin-compile-time": "^0.1.2",
+    "vite-plugin-ejs": "^1.4.3",
+    "vite-plugin-html-minifier-terser": "^2.0.0",
+    "vite-plugin-prismjs": "^0.0.8",
+    "vite-plugin-public-path": "^0.0.3",
     "wastyle": "^0.0.5",
-    "web-streams-polyfill": "^3.0.1",
-    "workerize-loader": "^1.3.0",
-    "xss": "^1.0.8"
+    "web-streams-polyfill": "^3.2.0",
+    "xss": "^1.0.10"
   },
-  "config-overrides-path": "react-app-rewired.js",
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "format": "prettier --write \"*.js\" \"scripts/**/*.js\" \"src/**/*.ts\" \"src/**/*.js\" \"src/**/*.tsx\"",
-    "check-style": "prettier --check \"*.js\" \"scripts/**/*.js\" \"src/**/*.ts\" \"src/**/*.js\" \"src/**/*.tsx\"",
+    "start": "vite",
+    "build": "vite build",
+    "format": "prettier --ignore-path .gitignore --write \"**/*.{js,ts,tsx}\"",
+    "check-style": "prettier --ignore-path .gitignore --check \"**/*.{js,ts,tsx}\"",
     "generate-api": "node scripts/generate-api.js $API_HOST/docs-json && prettier --write \"src/api-generated/**/*.ts\"",
     "postinstall": "patch-package"
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
   },
   "description": "The web frontend of Lyrio",
   "repository": "https://github.com/lyrio-dev/ui",
