@@ -1,5 +1,6 @@
-const fs = require("fs");
-const path = require("path");
+import fs from "fs";
+import path from "path";
+import url from "url";
 
 function escapeLocalizedMessage(text) {
   // Remove the "to-be-translated" prefix from message first
@@ -13,7 +14,11 @@ function escapeLocalizedMessage(text) {
   return text;
 }
 
-function escapeLocalizedMessages(object) {
+interface LocalizedMessages {
+  [key: string]: string | LocalizedMessages;
+}
+
+function escapeLocalizedMessages(object: string | LocalizedMessages) {
   if (typeof object === "string") return escapeLocalizedMessage(object);
 
   const result = {};
@@ -24,7 +29,7 @@ function escapeLocalizedMessages(object) {
 }
 
 // From https://stackoverflow.com/questions/5827612/node-js-fs-readdir-recursive-directory-search
-async function* readDirectoryRecursively(dir) {
+async function* readDirectoryRecursively(dir: string): AsyncGenerator<string> {
   const dirents = await fs.promises.readdir(dir, { withFileTypes: true });
   for (const dirent of dirents) {
     const res = path.resolve(dir, dirent.name);
@@ -36,29 +41,39 @@ async function* readDirectoryRecursively(dir) {
   }
 }
 
-const readMessageFile = async filename => {
+const readMessageFile = async (filename: string) => {
   try {
-    return new Function(await fs.promises.readFile(require.resolve(filename), "utf-8"))();
+    return new Function(await fs.promises.readFile(filename, "utf-8"))();
   } catch (e) {
     throw new Error(`\n  Error in ${filename}:\n  ${e.stack}`);
   }
 };
 
-module.exports = async (locale, loaderContext) => {
-  loaderContext.addDependency(__filename);
+/**
+ * @param fromUrl The `import.meta.url` in **compile time** of <some locale>/import.ts
+ */
+export default async (fromUrl: string) => {
+  const currentFile = url.fileURLToPath(fromUrl);
+  const currentDirectory = path.dirname(currentFile);
+
+  const watchFiles: string[] = [];
+  watchFiles.push(currentFile);
 
   const result = {};
-  const localeDirectory = path.resolve(__dirname, locale);
+  const localeDirectory = path.resolve(currentDirectory);
   for await (const absolutePath of readDirectoryRecursively(localeDirectory)) {
+    if (!absolutePath.endsWith(".js")) continue;
+
     const relativePath = path.relative(localeDirectory, absolutePath);
     const objectPath = relativePath.slice(0, -3); // Remove ".js"
 
-    loaderContext.addDependency(absolutePath);
+    watchFiles.push(absolutePath);
     result[objectPath.split("/").join(".")] = escapeLocalizedMessages(await readMessageFile(absolutePath));
   }
 
   return {
     cachable: true,
-    code: `module.exports = require("flat")(${JSON.stringify(result)});`
+    code: `import flat from "flat";export default flat(${JSON.stringify(result)});`,
+    watchFiles: watchFiles
   };
 };
