@@ -1,11 +1,7 @@
-// Public path is specfied in HTML and maybe dynamicly generated by server for different request.
-// See `window.publicPath` in ../public/index.html
-__webpack_public_path__ = window["publicPath"];
-
 /**
  * The initialization sequence:
  *
- * index.ts                  (webpack entry)
+ * index.ts                  (vite entry, await polyfill & xdomain)
  * => index.tsx              (app entry)
  *    => misc/webfonts.ts
  *    => initApp.ts          (session initialization)
@@ -16,6 +12,6 @@ __webpack_public_path__ = window["publicPath"];
  *          => Layout.tsx    (app view layout)
  *          => page routes
  */
-require("./index.tsx");
+import "./index.tsx";
 
 export {}; // Fix the error "All files must be modules when the '--isolatedModules' flag is provided".
