@@ -153,8 +153,7 @@ return {
     auto_add_testcases: {
       auto_add_testcases: "自动添加测试点",
       subtask: "子任务",
-      help:
-        "输入正则表达式以匹配文件名，使用捕获组来表示输入输出文件名中相同的部分。\n如果两个文件名分别被输入、输出的正则表达式匹配，且其每个捕获组的文本对应相同，则它们成为一组测试点。",
+      help: "输入正则表达式以匹配文件名，使用捕获组来表示输入输出文件名中相同的部分。\n如果两个文件名分别被输入、输出的正则表达式匹配，且其每个捕获组的文本对应相同，则它们成为一组测试点。",
       help_no_output: "输入正则表达式以匹配文件名，每个被匹配的文件会作为一个测试点的输入文件。",
       input_file: "输入文件",
       output_file: "输出文件",
