@@ -15,9 +15,9 @@ import {
   Ref
 } from "semantic-ui-react";
 import { v4 as uuid } from "uuid";
-import lodashIsEqual from "lodash.isequal";
+import lodash from "lodash";
 import { WritableStream } from "web-streams-polyfill/ponyfill/es6";
-import * as streamsaver from "streamsaver";
+import streamsaver from "streamsaver";
 import pAll from "p-all";
 import { useDebounce } from "use-debounce";
 
@@ -403,7 +403,7 @@ let FileTable: React.FC<FileTableProps> = props => {
       if (fileUuids.includes(fileUuid)) newSelectedFiles.add(fileUuid);
     }
 
-    if (!lodashIsEqual(selectedFiles, newSelectedFiles)) setSelectedFiles(newSelectedFiles);
+    if (!lodash.isEqual(selectedFiles, newSelectedFiles)) setSelectedFiles(newSelectedFiles);
   }, [props.files]);
 
   function onSelectAll(checked: boolean) {
